+++ src/snowflake/cli/plugins/spcs/compute_pool/commands.py
@@ -25,8 +25,10 @@ def _compute_pool_name_callback(name: str) -> str:
     """
     Verifies that compute pool name is a single valid identifier.
     """
-    if not is_valid_object_name(name, 0):
-        raise ClickException(f"'{name}' is not a valid compute pool name.")
+    if not is_valid_object_name(name, max_depth=0, allow_quoted=False):
+        raise ClickException(
+            f"'{name}' is not a valid compute pool name. Note that compute pool names must be unquoted identifiers."
+        )
     return name