+++ src/snowflake/cli/plugins/spcs/services/commands.py
@@ -30,8 +30,10 @@ app = SnowTyper(
 
 
 def _service_name_callback(name: str) -> str:
-    if not is_valid_object_name(name, 2):
-        raise ClickException(f"'{name}' is not a valid service name.")
+    if not is_valid_object_name(name, max_depth=2, allow_quoted=False):
+        raise ClickException(
+            f"'{name}' is not a valid service name. Note service names must be unquoted identifiers. The same constraint also applies to database and schema names where you create a service."
+        )
     return name