+++ tests/spcs/test_services.py
@@ -9,6 +9,8 @@ from snowflake.cli.plugins.spcs.services.manager import ServiceManager
 from tests.testing_utils.fixtures import *
 from snowflake.cli.api.project.util import to_string_literal
 from snowflake.cli.plugins.object.common import Tag
+from snowflake.cli.plugins.spcs.services.commands import _service_name_callback
+from click import ClickException
 
 SPEC_CONTENT = dedent(
     """
@@ -34,10 +36,8 @@ SPEC_DICT = {
 }
 
 
-@patch(
-    "snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_schema_query"
-)
-def test_create_service(mock_execute_schema_query, other_directory):
+@patch("snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_query")
+def test_create_service(mock_execute_query, other_directory):
     service_name = "test_service"
     compute_pool = "test_pool"
     min_instances = 42
@@ -55,7 +55,7 @@ def test_create_service(mock_execute_schema_query, other_directory):
     comment = "'user\\'s comment'"
 
     cursor = Mock(spec=SnowflakeCursor)
-    mock_execute_schema_query.return_value = cursor
+    mock_execute_query.return_value = cursor
 
     result = ServiceManager().create(
         service_name=service_name,
@@ -82,7 +82,7 @@ def test_create_service(mock_execute_schema_query, other_directory):
             "WITH TAG (test_tag='test value',key='value')",
         ]
     )
-    actual_query = " ".join(mock_execute_schema_query.mock_calls[0].args[0].split())
+    actual_query = " ".join(mock_execute_query.mock_calls[0].args[0].split())
     assert expected_query == actual_query
     assert result == cursor
 
@@ -154,7 +154,6 @@ def test_create_service_cli(mock_create, other_directory, runner):
         ]
     )
     assert result.exit_code == 0, result.output
-    print(mock_create.mock_calls[0])
     mock_create.assert_called_once_with(
         service_name="test_service",
         compute_pool="test_pool",
@@ -196,9 +195,7 @@ def test_create_service_with_invalid_spec(mock_read_yaml):
 
 
 @patch("snowflake.cli.plugins.spcs.services.manager.ServiceManager._read_yaml")
-@patch(
-    "snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_schema_query"
-)
+@patch("snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_query")
 @patch("snowflake.cli.plugins.spcs.services.manager.handle_object_already_exists")
 def test_create_repository_already_exists(mock_handle, mock_execute, mock_read_yaml):
     service_name = "test_object"
@@ -226,32 +223,39 @@ def test_create_repository_already_exists(mock_handle, mock_execute, mock_read_y
     )
 
 
-@patch(
-    "snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_schema_query"
-)
-def test_status(mock_execute_schema_query):
+@patch("snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_query")
+def test_status(mock_execute_query):
     service_name = "test_service"
     cursor = Mock(spec=SnowflakeCursor)
-    mock_execute_schema_query.return_value = cursor
+    mock_execute_query.return_value = cursor
     result = ServiceManager().status(service_name)
     expected_query = "CALL SYSTEM$GET_SERVICE_STATUS('test_service')"
-    mock_execute_schema_query.assert_called_once_with(expected_query)
+    mock_execute_query.assert_called_once_with(expected_query)
     assert result == cursor
 
 
-@patch(
-    "snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_schema_query"
-)
-def test_logs(mock_execute_schema_query):
+@patch("snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_query")
+def test_status_qualified_name(mock_execute_query):
+    service_name = "db.schema.test_service"
+    cursor = Mock(spec=SnowflakeCursor)
+    mock_execute_query.return_value = cursor
+    result = ServiceManager().status(service_name)
+    expected_query = f"CALL SYSTEM$GET_SERVICE_STATUS('{service_name}')"
+    mock_execute_query.assert_called_once_with(expected_query)
+    assert result == cursor
+
+
+@patch("snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_query")
+def test_logs(mock_execute_query):
     service_name = "test_service"
     container_name = "test_container"
     cursor = Mock(spec=SnowflakeCursor)
-    mock_execute_schema_query.return_value = cursor
+    mock_execute_query.return_value = cursor
     result = ServiceManager().logs(service_name, "10", container_name, 42)
     expected_query = (
         "call SYSTEM$GET_SERVICE_LOGS('test_service', '10', 'test_container', 42);"
     )
-    mock_execute_schema_query.assert_called_once_with(expected_query)
+    mock_execute_query.assert_called_once_with(expected_query)
     assert result == cursor
 
 
@@ -263,13 +267,11 @@ def test_read_yaml(other_directory):
     assert result == json.dumps(SPEC_DICT)
 
 
-@patch(
-    "snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_schema_query"
-)
-def test_upgrade_spec(mock_execute_schema_query, other_directory):
+@patch("snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_query")
+def test_upgrade_spec(mock_execute_query, other_directory):
     service_name = "test_service"
     cursor = Mock(spec=SnowflakeCursor)
-    mock_execute_schema_query.return_value = cursor
+    mock_execute_query.return_value = cursor
     tmp_dir = Path(other_directory)
     spec_path = tmp_dir / "spec.yml"
     spec_path.write_text(SPEC_CONTENT)
@@ -277,7 +279,7 @@ def test_upgrade_spec(mock_execute_schema_query, other_directory):
     expected_query = (
         f"alter service {service_name} from specification $$ {json.dumps(SPEC_DICT)} $$"
     )
-    actual_query = " ".join(mock_execute_schema_query.mock_calls[0].args[0].split())
+    actual_query = " ".join(mock_execute_query.mock_calls[0].args[0].split())
     assert expected_query == actual_query
     assert result == cursor
 
@@ -302,16 +304,14 @@ def test_upgrade_spec_cli(mock_upgrade_spec, mock_cursor, runner, other_director
     assert "Statement executed successfully" in result.output
 
 
-@patch(
-    "snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_schema_query"
-)
-def test_list_endpoints(mock_execute_schema_query):
+@patch("snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_query")
+def test_list_endpoints(mock_execute_query):
     service_name = "test_service"
     cursor = Mock(spec=SnowflakeCursor)
-    mock_execute_schema_query.return_value = cursor
+    mock_execute_query.return_value = cursor
     result = ServiceManager().list_endpoints(service_name)
     expected_query = f"show endpoints in service test_service"
-    mock_execute_schema_query.assert_called_once_with(expected_query)
+    mock_execute_query.assert_called_once_with(expected_query)
     assert result == cursor
 
 
@@ -330,16 +330,14 @@ def test_list_endpoints_cli(mock_list_endpoints, mock_cursor, runner):
     assert "test-snowflakecomputing.app" in result.output
 
 
-@patch(
-    "snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_schema_query"
-)
-def test_suspend(mock_execute_schema_query):
+@patch("snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_query")
+def test_suspend(mock_execute_query):
     service_name = "test_service"
     cursor = Mock(spec=SnowflakeCursor)
-    mock_execute_schema_query.return_value = cursor
+    mock_execute_query.return_value = cursor
     result = ServiceManager().suspend(service_name)
     expected_query = f"alter service {service_name} suspend"
-    mock_execute_schema_query.assert_called_once_with(expected_query)
+    mock_execute_query.assert_called_once_with(expected_query)
     assert result == cursor
 
 
@@ -355,16 +353,14 @@ def test_suspend_cli(mock_suspend, mock_cursor, runner):
     assert "Statement executed successfully" in result.output
 
 
-@patch(
-    "snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_schema_query"
-)
-def test_resume(mock_execute_schema_query):
+@patch("snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_query")
+def test_resume(mock_execute_query):
     service_name = "test_service"
     cursor = Mock(spec=SnowflakeCursor)
-    mock_execute_schema_query.return_value = cursor
+    mock_execute_query.return_value = cursor
     result = ServiceManager().resume(service_name)
     expected_query = f"alter service {service_name} resume"
-    mock_execute_schema_query.assert_called_once_with(expected_query)
+    mock_execute_query.assert_called_once_with(expected_query)
     assert result == cursor
 
 
@@ -380,10 +376,8 @@ def test_resume_cli(mock_resume, mock_cursor, runner):
     assert "Statement executed successfully" in result.output
 
 
-@patch(
-    "snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_schema_query"
-)
-def test_set_property(mock_execute_schema_query):
+@patch("snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_query")
+def test_set_property(mock_execute_query):
     service_name = "test_service"
     min_instances = 2
     max_instances = 3
@@ -391,7 +385,7 @@ def test_set_property(mock_execute_schema_query):
     auto_resume = False
     comment = to_string_literal("this is a test")
     cursor = Mock(spec=SnowflakeCursor)
-    mock_execute_schema_query.return_value = cursor
+    mock_execute_query.return_value = cursor
     result = ServiceManager().set_property(
         service_name=service_name,
         min_instances=min_instances,
@@ -410,7 +404,7 @@ def test_set_property(mock_execute_schema_query):
             f"comment = {comment}",
         ]
     )
-    mock_execute_schema_query.assert_called_once_with(expected_query)
+    mock_execute_query.assert_called_once_with(expected_query)
     assert result == cursor
 
 
@@ -482,9 +476,7 @@ def test_set_property_no_properties_cli(mock_set, runner):
     )
 
 
-@patch(
-    "snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_schema_query"
-)
+@patch("snowflake.cli.plugins.spcs.services.manager.ServiceManager._execute_query")
 def test_unset_property(mock_execute_query):
     service_name = "test_service"
     cursor = Mock(spec=SnowflakeCursor)
@@ -561,3 +553,29 @@ def test_unset_property_with_args(runner):
     )
     assert result.exit_code == 2, result.output
     assert "Got unexpected extra argument" in result.output
+
+
+def test_invalid_service_name(runner):
+    invalid_service_name = "account.db.schema.name"
+    result = runner.invoke(["spcs", "service", "status", invalid_service_name])
+    assert result.exit_code == 1
+    assert f"'{invalid_service_name}' is not a valid service name." in result.output
+
+
+@patch("snowflake.cli.plugins.spcs.services.commands.is_valid_object_name")
+def test_service_name_parser(mock_is_valid_object_name):
+    service_name = "db.schema.test_service"
+    mock_is_valid_object_name.return_value = True
+    assert _service_name_callback(service_name) == service_name
+    mock_is_valid_object_name.assert_called_once_with(
+        service_name, max_depth=2, allow_quoted=False
+    )
+
+
+@patch("snowflake.cli.plugins.spcs.services.commands.is_valid_object_name")
+def test_service_name_parser_invalid_object_name(mock_is_valid_object_name):
+    invalid_service_name = "account.db.schema.test_service"
+    mock_is_valid_object_name.return_value = False
+    with pytest.raises(ClickException) as e:
+        _service_name_callback(invalid_service_name)
+    assert f"'{invalid_service_name}' is not a valid service name." in e.value.message