+++ RELEASE-NOTES.md
@@ -5,7 +5,7 @@
 ## New additions
 
 ## Fixes and improvements
-* Updated help messages for `spcs` commands.
+
 
 # v2.1.0
 
@@ -45,7 +45,9 @@
 * Restricted permissions of automatically created files
 * Fixed bug where `spcs service create` would not throw error if service with specified name already exists.
 * Logging into the file by default (INFO level)
-
+* Added validation that service, compute pool, and image repository names are unqualified identifiers.
+* `spcs service` commands now accept qualified names.
+* Updated help messages for `spcs` commands.
 
 # v2.0.0